feat: add Main page
