+++ src/index.css
@@ -4,7 +4,6 @@
 }
 
 #root {
-  display: flex;
   margin: auto;
   background-color: #fff;
 }
